requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
certifi==2018.10.15
chardet==3.0.4
cloudpickle==0.6.1
configparser==3.5.0
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
enum34==1.1.6
futures==3.2.0
idna==2.7
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==2.2.3
mccabe==0.6.1
networkx==2.2
numpy==1.15.4
opencv-python==3.4.4.19
PassportEye==1.2.4
pdfminer==20140328
Pillow==5.3.0
pyasn1==0.4.4
pycrypto==2.6.1
pylint==1.9.3
pymaging==0.1
pymaging-png==0.1
pyparsing==2.3.0
pypassport==2.0
pyscard==1.9.7
pytesseract==0.2.5
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
qrcode==6.0
requests==2.20.1
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
socketIO-client==0.7.2
subprocess32==3.5.3
toolz==0.9.0
urllib3==1.24.1
websocket-client==0.54.0
wrapt==1.10.11
